Move verify context from vm into chain package
